setup: Replace 'distribution' with 'distro' to match layerindex change
The community requested that the value 'distros' be used instead of
'distributions' in the webapi.  We want to continue to match the names
and terminology used by the layer index, so make a global change.

The word 'distribution' is still used in both the local.conf.sample in a
descriptive message, as well as the --help text for '--distros'.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>