bin/setup.py: Move the environment-setup link to each run
Since the buildtools can be updated, we want to make sure the environment-setup
link is always correct.  We now always verify if the link is correct and if not
fix the link.

The link is also added to the .gitignore file, as we do not want to store this
in the project git repository.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>