layer_index.py: Change the way things are sorted
We need to ensure that we always have a sorted layerBranches so that the
various --list-* items print in the same order each time.  This can also
affect resolution if two items provide the same response.

We also need to ensure that the layerindex entries, and contents of the entries
are sorted as well.  Otherwise when we write out the datafiles the data can
remain the same, but the contents can change.  If the data is written into a
git repository this will cause unnecessary diffs.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>