setup.py: Skip project configuration when in mirror mode
In mirror mode, we only update the README file to a custom README-MIRROR
version.  We avoid creating other project files that have no use in a mirror.

Move the replacement setup in update_project into it's own function.
This will allow use to use the new copySample function in different locations.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>