setup.py: Add mirror specific index
When creating a local mirror (--mirror option), we need to also capture the
corresponding layer index entries for the items in that mirror.  We write out
those entries in a new 'mirror-index.git' repository that has been created
for the project.  The serialized index pieces must also include the layerBranch
and layerItems for any required dependencies, otherwise we will be unable to
resolve them in a partial (not --all-layers) mirror.

When we go to use the mirror, we need to first check if there may be a mirror
index available, and then fetch a local copy if there is.  We use the
git ls-remote functionality to see if there is a corresponding mirror index
for the branch we are working on.

When parsing the layer_index items, we first load the contents of the cache
and then skip any loading any indexes that had been loaded by the cached
contents.

This code does support caching multiple remote indexes, as well as having multiple
local indexes with only some being filled in by a cache.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>