settings.py: Change the default Wind River Linux config
Include the wr-fixes layers in all configurations.  This along with wr-base
provides the foundation for Wind River Linux.

Change the default distribution to 'wrlinux-small'.  This distribution
enables busybox support.  Most users seem to start with the minimal/small
image targets, and without busybox being enabled by default this is causing
problems.

Also expand on the project setup notes to include the common Wind River
images recipes with information as to which distro they should use.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>