setup.sh: Fix REPO_URL processing
The REPO_URL needs to be added by the wrl_repo.sh script, however the function
was not being run without an eval.

We also adjust the REPO_URL to base in the base directory of the base git
repository, as this is expected by the repo tool when mirroring.

The repo tool also automatically mirrors a copy of itself, so remove it from
the default.xml file.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>