wrl_buildtools.sh: Change update logic
Instead of doing the update logic based on the URL, we do it based on the
repository date code.  If the date code has not changed, we do not update.

We also move the buildtools.git cache into the project itself.  This
simplifies the program and doesn't require any additional space.

Finally we allow environment settings to override the defaults.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>