data/xml: references must not end in .git
Repositories urls ending in .git will break when mirrored.  If you reference
a repository called foo.git, when mirrored it will become foo.git.git.

When git tries to pull down 'foo.git.git' it will be unable to due to the
duplicate '.git' suffix.

We adjust all of the 'name' fields and remove '.git', but they need to remain
in the 'path' fields.  (git will automatically try the 'name' and the
'name'+.git, so everything keeps working as expected.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>

fix data/xml/wr-kernel.xml

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>