layer_index.py: Fix mirror generation vcs_url replacement
The current code ended up only modifying the scheme://host which resulted in
the url building up creating invalid urls.

We want to use base_url as the replacement if the vcs_url starts with that
value, otherwise we fall back to using the scheme://host to map to the auto
generate 'remote'.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>