windshare: Initial windshare integration implementation.
Windshare contains a series of 'folders' that are tied to customer
entitlements.

The system discovers the folders the user is entitled to by loading a url
that containss a json file listing active folders.

Each folder is then processed looking for a mirror-index repository.  Each
mirror-index is loaded and into the local mirror-index cache.  After they
are all available, we process them -- adding folder info to the vcs_url, and
various XML files.  The results are then written out into the branch a
normal install from a mirror-index would have used.

All of the other processing happens as expected.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>