data/environment.d: Implement username and password settings
When connecting to a password protected server, we need a way to enter the
credentials once and store them.

The new setup_askpass is a small python program that asks as an ASKPASS
interface to request the credentials from the user in order to get to the
initial online steps.  The credentials are only stored in memory and never
written to the disk.  This should limit any exposure.

The setup_anspass is a generic script to help us setup and control the
enablement of 'anspass'.  It also handles the transition from the python
askpass implementation to the more secure anspass.  For more information
on anspass see: https://github.com/WindRiver-OpenSourceLabs/anspass

The remaining files handle the load order and related processing:

1 - Check if we have buildtools, if so load them early.  We want
    anspass support ASAP.
2 - Check if we have anspass available, if so enable it.  (We
    assume at this point, if we can enable it -- it has been
    properly configured by the user.)
3 - If anspass is not available, load the python askpass
4 - Properly load the buildtools (download, update, etc as necessary)
5 - Setup anspass (if not already running) and transition askpass->anspass
    if necessary.
6 - Find the location of the Wind River git-repo, set REPO_URL

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>