utils_setup.py: Support login credentials for json access
Remove all references to urllib2 -- this does not exist in python3.

Refactor all of the urllib calls into a single util_setup.py provided
function: fetch_url(url, auth)

The new function will attempt to contact the URL, if an exception is thrown
(401) the system will attempt to authenticate using the GIT_ASKPASS, failing
that it will go to SSH_ASKPASS.  If neither are available it will throw
an exception.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>