Refactor the username/password setup, including askpass/anspass
The new method tries to reduce the work being performed.

The new approach is:

If the user has the anspass DB setup, we startup and use it.  This makes the
assumption that if the DB is there, there are credentials to use.

If anspass is not started, we will startup askpass to gather credentials.

After buildtools, we check if the user passed in the --user parameter.  This
indicates that they want a non-interactive install and are trying to pass
credentials.  So we startup anspass here.

<perform installations>

During shutdown processing, if askpass has any added credentials, we will
process them and add them to the anspass database.  Otherwise we just stop.

In the end what this does is:

If the user connects somewhere with no credentials, anspass is never started.

If the user connects where credentials are needed, they will be prompted for
them by askpass, and the credentials will be transferred to anspass after the
setup.sh actions are complete.

If the user has previously been asked for credentials, we'll just re-use them
and avoid prompting them.

If the user specifies credentials, we'll just use them and avoid prompting.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>