04_wrl_buildtools.sh: uname -p is not portable, use uname -m
Issue: LIN9-2615

We want to use uname -m, as it is a portable way to find the host arch.

If the uname -m result is invalid for some reason, we now provide a way
for the user to specify the SDKARCH in the environment.

Additionally, we will list in the error messages, all of the architectures
that we support.  (This is based on the contents of the buildtools archive.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>