bin/setup.py: fix --all-layers handle dl layers
Issue: LIN10-14

Fixed 1:
$ ./wrlinux-x/setup.sh --machines=qemux86 --distros=wrlinux-std-sato --all-layers

There are no foo-dl layers fetched, but the code said they should be.

Fixed 2: (on master branch)
$ ./wrlinux-x/setup.sh --machines=qemux86 --distros=wrlinux-std-sato --all-layers --dl-layers

The foo-dl-2.2 are fetched, this is wrong, it should fetch foo-dl-2.3
since it is on master branch, this is because the code didn't repect the
layer branch dependencies, this patch can fix the problem.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>