Add additional debug logging for fetch_url
Issue: LIN9-3597

If fetch_url is passed with a logging debuglevel, this level is then passed
to urllib.  Unfortunately urllib will only log to sys.stdout, so something
else in the system must have already capture this so it can be logged.

In addition, even in the case of an error, we will out the retrieved contents
as it might be useful to diagnose a URL failure.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>