windshare.py: Minor cleanup
Move the import sys to be global to the file.  This will make it easier
in future changes to call sys.exit(1) if needed.

Cleanup the generic exception handling to also include the type of the
exception.  Some exceptions, when printed, do not include the type and will
make this more difficult to determine how best to resolve the problem.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>