layer_index.py: Cleanup error handling when unable to access the index
Not being able to access the index is NOT fatal, as we will fall back to
the cached version of the index.  However, we need to let the user know that
we were not able to contact the index.  So make this a warning.

also fix minor errors in this exception handler.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>