windshare.py: Adjust handling of URLError exception when trying windshare
If the URL includes the word "windshare" we assume any URLErrors are fatal.

The most common type is authentication failure or DNS lookup failure, but
other failures are possible as well, including 404, 500, etc.

Non-windshare URLs are not fatal and only attempted.  This allows for a user
to publish a copy of the .zip or .udf file download from windshare onto
their own servers.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>