utils_setup.py: Switch from using the HTTP to HTTPSHandler by default
Issue: LIN9-3597

The HTTP and HTTPS handlers are generally compatible and will switch behaviors
between each other.  However if you start with the HTTP Handler and perform
an https connection, you will be unable to debug it.  Starting with the
HTTPSHandler ensures that you can debug both http and https connections.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>