setup.sh: Remove --kernel option
The kernel option was confusing for customers.  The option did no validation,
did not always take affect, and still required manual intervention in some
configurations.

Remove the option and add a message if a user enters it, instructing them it
is no longer supported and that they must manually change their local.conf
file.

Update the local.conf file as well to better explain to the user how to
adjust the setting, if they feel it is necessary.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>