flatten_mirror.py: Adjust the git-repo location and copying
Issue: LIN10-4070

When mirroring for our internal build, git-repo is at the base of the
project.  However, when mirroring from an external source (such as
Windshare), it may be located within a directory.   Change the rule to
look for the git-repo repository in the same directory as openembedded-core
and bitbake.

Also a small tweak to a log message, which hides a scary looking error
message -- that isn't an error!

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>