setup.sh: Check the BASEURL is valid before allowing the user to continue
Issue: LIN10-4081

The BASEURL is usually derived from the location that wrlinux-x was cloned from.
This should either be the path to a local directory on the disk or a properly
formatted URL.  (Note: the path may include relative pieces, that is fine,
but it's not allowed that the WHOLE path be relative, i.e. not start with '/'.)

Using 'ssh' syntax of 'user@host:path' is also not supported.

Valid BASEURL syntax includes:

   /path
   file:///path
   http://host/path
   https://host/path
   git://host/path
   ssh://user@host/path

In order to permit the user to override the BASEURL (vs having to reclone
wrlinux-x with a different URL), the argument parsing is now done twice.  Once
with the initial options, and then again with the environment.d extended
options.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>