setup_askpass: Enable askpass with SSH urls
Issue: LIN10-4081

The system has two password managers.  A memory resident 'askpass' and
the storage based 'anspass'.  The former is typically only used when
anspass is not available.

anspass has a limitation in the form of the question being asked, it
does not understand the 'question' sent by SSH.  So we keep anspass
functionality disabled in ssh is being used.  However we can now
enable 'askpass' functionality, as it does not have an issue with the
format of the question.

The user will be prompted for a password, as needed.  It is recommended
that the user use key authentication if possible.  However, if they must
define a password to be used they can use the environment WINDSHARE_PASS
or the --password argument.  Note, the --password argument can leak their
credentials via 'ps' to other users on the system.

setup_ssh: SSH will only ask, via SSH_ASKPASS, when it does not have access
to a tty and the DISPLAY is not set.  We force this behavior by adding a
wrapper "setup_ssh".  This clears the DISPLAY and uses setsid to also clear
the terminal.

03_wrl_askpass.sh: Update to configure the askpass system to have the ssh
password response, if predefined by the user.  Set GIT_SSH to use the new
setup_ssh wrapper.

setup_askpass: Unfortunatly the python getpass.getpass(prompt) does not
work when run when DISPLAY= and setsid has been used.  Since we're trying
to run this on the original pty, and not the modified tty -- this doesn't
make any sense.  But being unable to fix this, we need to implement our
own getpass functions.

We do this by storing the current termios settings, and then modifying it
to remove 'echo' and 'signal' processing.  We can then display the problem
and use readline for the standard (non-echo) processing.  As a precaution,
we do verify that the echo was properly set and warn the user if it was
not.  (The signal is blocked, otherwise a user who hits sigint/ctrl-c will
end up with a broken terminal.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>