anspass/askpass: Disable both when URL is ssh
Issue: LIN10-4081

SSH will always ask the password question if it needs it, it does not use
SSH_ASKPASS, unless DISPLAY is empty and there is no controlling
terminal.

The user should have an authorized_keys file setup on the server, otherwise
the numerous prompts will be difficult to deal with.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>