bin/utils_setup.py: Clean PYTHONHOME from the environment for Popen
Issue: LIN1018-2936

The username and password prompting may call out to an external program
that eventually runs a host version of python.  If this happens, the
PYTHONHOME setting from the buildtools wrapper may cause a failure.

Typical result is an error similar to:

   ImportError: No module named site

This error comes from the external program that was invoked by setup.py.

Ensure this is cleared.  This was missed in the previous commit for the
related issue LIN1018-2934.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>