logger_setup.py: Add isatty() to class LoggerOut()
Issue: LIN1019-1408

The LoggerOut() is used for replacing sys.stdout and sys.stderr, but other code
may use sys.stdout.isatty() to check wither the output is a tty or not, for
example, the logger needs to check and determine whether to enable the color or
not. so add isatty() to LoggerOut() to make it work.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>