Remove integrated git-repo, and dynamically download git-repo
Issue: LINUXEXEC-2866

Remove local copy of bin/repo

Add two new arguments: --repo-url and --repo-branch

These arguments will allow the user to specify an alternative location to
download git-repo.  These arguments are also passed to each of the invocations
of git repo.  git-repo is downloaded into bin/git-repo

The setup environment script, 06_wrl_repo is extended to identify not only
the url for git-repo, but also the branch (or tag) to use.

The new code will check for the passed in branch, if not will try the
BASEBRANCH, and if that doesn't exist will use the 'repo_branch_fallback'.

Currently repo_branch_fallback is 'master-wr'.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>