sanity: Adjust a few minor things for OOBE
Extend sanity.check_hosttools to allow additional tools to be passed in.
- These are checked for duplicates
- Similarly adjust the stand alone sanity.py to allow sys.argv to be passed
  to check_hosttools.
- Set sanity.py to executable

setup.py: Move the sanity check until after some other diagnostics, but before
the check to see if they failed.  This will allow an output with the max
number of error messages possible to allow a user to correct any problems
without having to re-run the program multiple times.

Additionally, pass in the self.tool_list to the sanity.check_hosttools call
as it provides a better error message that what the existing get_path is
capable of doing.

Finally adjust get_path to NOT error, but add a None value when it can't
find a tool.  This allows all of the errors to be collected at once if
multiple items are missing.  (Note, this should never trigger because we
already checked for existence in the earlier call to check_hosttools.)

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>