bin/setup.py: Check for empty index, and do a clean error if found
Issue: LINUXEXEC-3105

If the index can't load, instead of reporting a nice clear error, the system
would give the user a backtrace and the cause of the issue may be obscured.

Resolve this by checking if at least one database has been added to the list.
Otherwise provide a critical failure and proper error message and exit.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>