setup.sh: clearly require python2
Issue: LIN1019-3194

For OOBE, if users are trying to setup wrlinux in a newly installed host
such Ubuntu18.04, they will meet the following error because python2 is
not installed by default.

  Stopping: an error occurred in askpass_setup.
  /usr/bin/env: 'python': No such file or directory

03_wrl_askpass.sh uses setup_askpass which is written in python2. And this
script is executed before setup.py which checks host tools. So we need to
give out clear message to user that python2 is required for setup to run
as expected.

And in the long run, we should convert askpass_setup to python3.

Signed-off-by: Chen Qi <Qi.Chen@windriver.com>