bin/setup.py: Checkout to a named repo branch
Issue: LINCD-1376

The repo command didn't checkout out to a specific branch, this caused
confusions when switched from branches, for example, when re-ran setup from
master-wr to WRLINUX_CI, repo's current branch was still master-wr (named
default), this was incorrect, now checkout to a named repo branch to fix this
problem.

E.g.:
$ cd .repo/repo
$ git branch
* master-wr

It was 'default' in the past. When re-run setup with WRLINUX_CI, it will be
updated to WRLINUX_CI as well.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>