environment.d/06_wrl_repo.sh: Drop bin/.git-repo-rev
Issue: LINCD-1376

It was used for saving current repo branch and would be used when re-run setup,
this caused problems when switching branches, for example, when re-run setup
from master-wr to WRLINUX_CI, we need the WRLINUX_CI branch, but not master-wr.
So drop it, and will get repo branch from arguments or repo_branch_fallback.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>