setup: add option --buildtools-type
Issue: LINUXEXEC-6868
Issue: LINCD-1378

Add an option '--buildtools-type' for setup. It then could select which
buildtools will be deploy, either 'buildtools-tarball' or
'buildtools-extended-tarball'. And the former is default.

It also checks host gcc version. If it is older than version 5.0, then
force to use buildtools extended to support c++14 standard.

Signed-off-by: Kai Kang <kai.kang@windriver.com>