Enable handling of local and remote layers in setup
The --layers option now accepts a path to a layer on the local
filesystem or a remote url. The local paths are added to the
bblayers.conf.sample. The remote layers are added to the git-repo
manifest.xml so they are downloaded by git-repo. If the remote
repository name matches a layer from the layerindex, the remote layer
and branch will override the layerindex layer.

Signed-off-by: Konrad Scherer <Konrad.Scherer@windriver.com>