setup_anspass: check the credentials in anspass database
Issue: LINUXEXEC-10887

Previously, the anspass database always exists in an project
which has run setup program before. It's annoying when user has
changed the credentials and has no chance to reinput the new one
though the old one in anspass database is already incorrect.

Check the credentials first and let user choose to remove the
anspass database to reinput the new one if the old credentials
in anspass database not correct.

Signed-off-by: Mingli Yu <mingli.yu@windriver.com>