data/environment.d/04_wrl_buildtools.sh: Update buildtools to 10.20.52.0
Issue: LINUXEXEC-11636

The ccahe has been upgraded to 4.1 which depends on cmake, this makes it hard
to use ccache-native, and oe-core suggests using the one from host, so we add
nativesdk-ccache to buildtools to make sure all the recipes can be built.
Otherwise, the recipes maybe failed to build with different versions' ccache.

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>